chore: move test_helpers.go to testutil package
