+++ app/testutil/test_helpers.go
@@ -1,10 +1,11 @@
-package app
+package testutil
 
 import (
 	"testing"
 	"time"
 
 	"github.com/stretchr/testify/require"
+	"github.com/tendermint/tendermint/crypto/ed25519"
 	dbm "github.com/tendermint/tm-db"
 
 	abci "github.com/tendermint/tendermint/abci/types"
@@ -22,6 +23,8 @@ import (
 	authtypes "github.com/cosmos/cosmos-sdk/x/auth/types"
 	banktypes "github.com/cosmos/cosmos-sdk/x/bank/types"
 	stakingtypes "github.com/cosmos/cosmos-sdk/x/staking/types"
+
+	"github.com/regen-network/regen-ledger/v4/app"
 )
 
 // DefaultConsensusParams defines the default Tendermint consensus params used in
@@ -50,7 +53,7 @@ type SetupOptions struct {
 	InvCheckPeriod     uint
 	HomePath           string
 	SkipUpgradeHeights map[int64]bool
-	EncConfig          EncodingConfig
+	EncConfig          app.EncodingConfig
 	AppOpts            types.AppOptions
 }
 
@@ -59,15 +62,15 @@ func DefaultOptions() SetupOptions {
 		Logger:             log.NewNopLogger(),
 		DB:                 dbm.NewMemDB(),
 		InvCheckPeriod:     5,
-		HomePath:           DefaultNodeHome,
+		HomePath:           app.DefaultNodeHome,
 		SkipUpgradeHeights: map[int64]bool{},
-		EncConfig:          MakeEncodingConfig(),
+		EncConfig:          app.MakeEncodingConfig(),
 		AppOpts:            EmptyAppOptions{},
 	}
 }
 
 // NewAppWithCustomOptions initializes a new RegenApp with custom options.
-func NewAppWithCustomOptions(t *testing.T, isCheckTx bool, options SetupOptions) *RegenApp {
+func NewAppWithCustomOptions(t *testing.T, isCheckTx bool, options SetupOptions) *app.RegenApp {
 	t.Helper()
 
 	privVal := mock.NewPV()
@@ -85,9 +88,9 @@ func NewAppWithCustomOptions(t *testing.T, isCheckTx bool, options SetupOptions)
 		Coins:   sdk.NewCoins(sdk.NewCoin(sdk.DefaultBondDenom, sdk.NewInt(100000000000000))),
 	}
 
-	app := NewRegenApp(options.Logger, options.DB, nil, true, options.SkipUpgradeHeights, options.HomePath, options.InvCheckPeriod, options.EncConfig, options.AppOpts)
-	genesisState := NewDefaultGenesisState(app.appCodec)
-	genesisState = genesisStateWithValSet(t, app, genesisState, valSet, []authtypes.GenesisAccount{acc}, balance)
+	regenApp := app.NewRegenApp(options.Logger, options.DB, nil, true, options.SkipUpgradeHeights, options.HomePath, options.InvCheckPeriod, options.EncConfig, options.AppOpts)
+	genesisState := app.NewDefaultGenesisState(regenApp.AppCodec())
+	genesisState = genesisStateWithValSet(t, regenApp, genesisState, valSet, []authtypes.GenesisAccount{acc}, balance)
 
 	if !isCheckTx {
 		// init chain must be called to stop deliverState from being nil
@@ -95,7 +98,7 @@ func NewAppWithCustomOptions(t *testing.T, isCheckTx bool, options SetupOptions)
 		require.NoError(t, err)
 
 		// Initialize the chain
-		app.InitChain(
+		regenApp.InitChain(
 			abci.RequestInitChain{
 				Validators:      []abci.ValidatorUpdate{},
 				ConsensusParams: DefaultConsensusParams,
@@ -104,14 +107,14 @@ func NewAppWithCustomOptions(t *testing.T, isCheckTx bool, options SetupOptions)
 		)
 	}
 
-	return app
+	return regenApp
 }
 
 func genesisStateWithValSet(t *testing.T,
-	app *RegenApp, genesisState GenesisState,
+	app *app.RegenApp, genesisState app.GenesisState,
 	valSet *tmtypes.ValidatorSet, genAccs []authtypes.GenesisAccount,
 	balances ...banktypes.Balance,
-) GenesisState {
+) app.GenesisState {
 	// set genesis accounts
 	authGenesis := authtypes.NewGenesisState(authtypes.DefaultParams(), genAccs)
 	genesisState[authtypes.ModuleName] = app.AppCodec().MustMarshalJSON(authGenesis)
@@ -177,4 +180,15 @@ type EmptyAppOptions struct{}
 // Get implements AppOptions
 func (ao EmptyAppOptions) Get(o string) interface{} {
 	return nil
+}
+
+// CreateRandomAccounts is a function return a list of randomly generated AccAddresses
+func CreateRandomAccounts(numAccts int) []sdk.AccAddress {
+	testAddrs := make([]sdk.AccAddress, numAccts)
+	for i := 0; i < numAccts; i++ {
+		pk := ed25519.GenPrivKey().PubKey()
+		testAddrs[i] = sdk.AccAddress(pk.Address())
+	}
+
+	return testAddrs
 }
